Struct safe_arch::m128

#[repr(transparent)]pub struct m128(pub __m128);

The data for a 128-bit SSE register of four f32 lanes.

This is very similar to having [f32; 4]. The main difference is that it's aligned to 16 instead of just 4, and of course you can perform various intrinsic operations on it.

Implementations

impl m128

#[must_use]pub fn to_array(self) -> [f32; 4]

Transmutes the m128 to an array.

Same as m.into(), just lets you be more explicit about what's happening.

#[must_use]pub fn from_array(f: [f32; 4]) -> Self

Transmutes an array into m128.

Same as m128::from(arr), it just lets you be more explicit about what's happening.

#[must_use]pub fn to_bits(self) -> [u32; 4]

Converts into the bit patterns of these floats ([u32;4]).

Like f32::to_bits, but all four lanes at once.

#[must_use]pub fn from_bits(bits: [u32; 4]) -> Self

Converts from the bit patterns of these floats ([u32;4]).

Like f32::from_bits, but all four lanes at once.