Struct safe_arch::m128i

#[repr(transparent)]pub struct m128i(pub __m128i);

The data for a 128-bit SSE register of integer data.

The exact layout to view the type as depends on the operation used.

From and Into impls are provided for all the relevant signed integer array types.

Formatting impls print as four i32 values just because they have to pick something. If you want an alternative you can turn it into an array and print as you like.