Struct safe_arch::m128d

#[repr(transparent)]pub struct m128d(pub __m128d);

The data for a 128-bit SSE register of two f64 values.

This is very similar to having [f64; 2]. The main difference is that it's aligned to 16 instead of just 4, and of course you can perform various intrinsic operations on it.

Implementations

impl m128d

#[must_use]pub fn to_array(self) -> [f64; 2]

Transmutes the m128d to an array.

Same as m.into(), just lets you be more explicit about what's happening.

#[must_use]pub fn from_array(f: [f64; 2]) -> Self

Transmutes an array into m128d.

Same as m128d::from(arr), it just lets you be more explicit about what's happening.

#[must_use]pub fn to_bits(self) -> [u64; 2]

Converts into the bit patterns of these doubles ([u64;2]).

Like f64::to_bits, but both lanes at once.

#[must_use]pub fn from_bits(bits: [u64; 2]) -> Self

Converts from the bit patterns of these doubles ([u64;2]).

Like f64::from_bits, but both lanes at once.