Struct safe_arch::m256

#[repr(transparent)]pub struct m256(pub __m256);

The data for a 256-bit AVX register of eight f32 lanes.

This is very similar to having [f32; 8]. The main difference is that it's aligned to 32 instead of just 4, and of course you can perform various intrinsic operations on it.

Implementations

impl m256

#[must_use]pub fn to_array(self) -> [f32; 8]

Transmutes the m256 to an array.

Same as m.into(), just lets you be more explicit about what's happening.

#[must_use]pub fn from_array(f: [f32; 8]) -> Self

Transmutes an array into m256.

Same as m256::from(arr), it just lets you be more explicit about what's happening.

#[must_use]pub fn to_bits(self) -> [u32; 8]

Converts into the bit patterns of these floats ([u32;8]).

Like f32::to_bits, but all eight lanes at once.

#[must_use]pub fn from_bits(bits: [u32; 8]) -> Self

Converts from the bit patterns of these floats ([u32;8]).

Like f32::from_bits, but all eight lanes at once.